feat: GH-29 Switch to gradle typesafe accessors        Add movie detail screen        Add screenshot tests
