+++ core/ui/src/main/kotlin/me/jerryokafor/core/ui/components/MoviePoster.kt
@@ -53,9 +53,9 @@ import me.jerryokafor.core.ds.theme.IheNkiriTheme
 const val MOVIE_POSTER_TEST_TAG = "movie_poster"
 private const val SIZE_ASPECT_RATIO = 0.8F
 
+@ExcludeFromJacocoGeneratedReport
 @ThemePreviews
 @Composable
-@ExcludeFromJacocoGeneratedReport
 fun PosterPreview() {
     IheNkiriTheme {
         Column(
