+++ core/ui/src/test/screenshots/Background/Background_dark_dynamic.png
