+++ core/ui/src/test/screenshots/Background/Background_dark_notDynamic.png
