+++ core/ui/src/test/screenshots/Background/Background_light_dynamic.png
