+++ core/ui/src/test/screenshots/Background/Background_light_notDynamic.png
