+++ core/ui/src/test/screenshots/GenreChip/GenreChip_dark_notDynamic.png
