+++ core/ui/src/test/screenshots/GenreChip/GenreChip_light_dynamic.png
