+++ core/ui/src/test/screenshots/MoviePoster/MoviePoster_dark_dynamic.png
