+++ core/ui/src/test/screenshots/MoviePoster/MoviePoster_dark_notDynamic.png
