+++ core/ui/src/test/screenshots/MoviePoster/MoviePoster_light_dynamic.png
