+++ core/ui/src/test/screenshots/MoviePoster/MoviePoster_light_notDynamic.png
