+++ core/ui/src/test/screenshots/MovieRating/MovieRating_dark_dynamic.png
