+++ core/ui/src/test/screenshots/MovieRating/MovieRating_light_dynamic.png
