+++ core/ui/src/test/screenshots/TrailerButton/TrailerButton_dark_notDynamic.png
