+++ core/ui/src/test/screenshots/TrailerButton/TrailerButton_light_dynamic.png
