+++ core/ui/src/test/screenshots/TrailerButton/TrailerButton_light_notDynamic.png
